Integrated Windows-specific build fixes from e250f17a:
- CMakeLists.txt: MinGW support (system SDL2 via find_package, SDL_MAIN_HANDLED,
  dbghelp for MSVC), sdl2.cmake find_package for non-MSVC Windows
- Crash handler split: crashhandler_windows.cpp/h, cross-platform refactor
- uae_server_thread.cpp: Full refactor with Windows-compatible patterns
- uae_vm_imp.cpp, vm_player_selector: Windows compatibility adjustments
- EASTL, qd, amDebugger: MSVC warning fixes (char_traits, tuple, Color, float)
- hardfile_host.cpp, sound.cpp, sysconfig.h, dummy.cpp: Windows I/O fixes
- compile_options.cmake: MSVC /W4 /WX enforcement

Conflicts resolved (5 files):
- adf_file.c, adf_file_util.h, adf_raw.c: Removed NDEBUG entirely (render-fix)
- calc.cpp: Took render-fix version, then fixed 5 new warnings:
  - 2x pointer-bool: vals→vals[0]
  - 3x constant-conversion: cast '| 0x80' to TCHAR
- uae_server_thread.cpp: Kept dbg-dev ordering (SDL_AtomicIncRef before unlock)

+++ libs/amDebugger/src/amDebugger/os/os_introspector.cpp
@@ -28,7 +28,7 @@ std::string OsIntrospector::readCString(uint32_t addr, size_t maxLen) {
     result.reserve(32);
     for (size_t i = 0; i < maxLen; ++i) {
         // Checking bounds could be done by getRealAddr but here we rely on getU8 returning 0 or safe data.
-        uint8_t c = m_vm->mem->getU8(addr + i);
+        uint8_t c = m_vm->mem->getU8(static_cast<uint32_t>(addr + i));
         if (c == 0) break;
         result.push_back((char)c);
     }
@@ -172,7 +172,7 @@ std::vector<LibraryInfo> OsIntrospector::scanLibraries() {
                 uint16_t opcode = readU16(lvoAddr);
                 if (opcode == 0x4EF9) { // JMP abs.L
                     LvoEntry lvo;
-                    lvo.offset = -offset;
+                    lvo.offset = static_cast<int16_t>(-offset);
                     lvo.isJump = true;
                     lvo.targetAddress = readU32(lvoAddr + 2);
                     lvo.funcName = getLvoName(lib.name, offset);
